SUCCESS
The tooltips problem has been resolved.

Because AJAX are run asynch, sometimes the parseCorp method was being
called before the parseChar method. Since parseChar is modifying the
HTML, it was resetting the tooltips.

A simple shuffle of method calls has fixed it. Corp is now ALWAYS called
after char is done, and tooltips are now always appearing.

Cleaned up some var names, the way jquery is interacting, and removed a
useless method.

Phew.